Large refactor of math expressions
Instead of keeping both lorentz and color tensors hardcoded in the
math expression enum, allow for a type argument to specify the tensor.
This makes the previously undefined case of math and color tensors in
a single expression illegal and makes it much easier to seperate spin
and color code from the math_expr module, which is now the helper module
it was always supposed to be.