Add a new TensorIndex trait for normalizing indices
This is necessary, since the distinction between triplet indices and
antitriplet indices for color means that it's much harder to discover
if a tensor actually contains a given summed index.
Additionally, the index matching by the parser has been switched to
consider all indices with the same underlying number as contracted,
compared to actually identical indices as before.
This adds the requirement that all index numbers must be unique, not
just unique for each index type. This is not a problem, since that was
a requirement for UFO anyway.